pyproject.toml
[project]
name = "ocreceipt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "groq>=0.33.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "supabase>=2.24.0",
]
